refactor(groups): extract discovery methods into GroupsDiscoveryService
Moves listMyGroups and searchGroups out of GroupsService (~437 lines)
into a new GroupsDiscoveryService, keeping GroupsService focused on
group CRUD (createGroup, findById, getGroup, updateGroup, deleteGroup).

GroupsController now injects both services. Spec files split accordingly.
