+++ apps/api/src/modules/groups/groups.module.ts
@@ -1,10 +1,11 @@
 import { Module } from '@nestjs/common';
 import { GroupsController } from './groups.controller';
 import { GroupsService } from './groups.service';
+import { GroupsDiscoveryService } from './groups-discovery.service';
 
 @Module({
   controllers: [GroupsController],
-  providers: [GroupsService],
+  providers: [GroupsService, GroupsDiscoveryService],
   exports: [GroupsService],
 })
 export class GroupsModule {}
