+++ apps/api/src/modules/groups/groups.service.spec.ts
@@ -14,7 +14,6 @@ import { CloudStorageService } from '@/modules/cloud-storage/cloud-storage.servi
 import { GroupsService } from './groups.service';
 import type { CreateGroupDto } from './dto/create-group.dto';
 import type { UpdateGroupDto } from './dto/update-group.dto';
-import type { SearchGroupsQueryDto } from './dto/search-groups-query.dto';
 import type { AuthenticatedUser } from '@/types/express';
 
 jest.mock('@google-cloud/storage', () => ({
@@ -335,27 +334,6 @@ describe('GroupsService', () => {
     });
   });
 
-  describe('listMyGroups', () => {
-    it('returns empty array when user has no active memberships', async () => {
-      mockGroupMembersFindMany.mockResolvedValue([]);
-
-      const result = await service.listMyGroups('user-uuid');
-
-      expect(result).toEqual([]);
-    });
-
-    it('returns groups for active memberships', async () => {
-      mockGroupMembersFindMany.mockResolvedValue([mockOwnerMembership]);
-      mockGroupsFindMany.mockResolvedValue([mockGroupRow]);
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow]);
-
-      const result = await service.listMyGroups('user-uuid');
-
-      expect(result).toHaveLength(1);
-      expect(result[0]?.id).toBe('group-uuid');
-    });
-  });
-
   describe('updateGroup', () => {
     const updateDto: UpdateGroupDto = { name: 'Updated Crew' };
 
@@ -650,235 +628,6 @@ describe('GroupsService', () => {
       mockCloudStorageDelete.mockRejectedValue(new Error('GCS unavailable'));
 
       await expect(service.deleteGroup(mockUser, 'group-uuid')).resolves.toBeUndefined();
-    });
-  });
-
-  describe('searchGroups', () => {
-    const baseQuery: SearchGroupsQueryDto = { limit: 20, offset: 0 };
-
-    const mockGroupRow2 = {
-      id: 'group-uuid-2',
-      name: 'Beach Crew',
-      description: 'Sun and surf',
-      cover: 'asset-uuid-2',
-      visibility: GroupVisibility.PUBLIC,
-      createdBy: 'other-user-uuid',
-      createdAt: new Date('2026-01-02T00:00:00.000Z'),
-      updatedAt: new Date('2026-01-02T00:00:00.000Z'),
-    };
-
-    const mockCoverAssetRow2 = { ...mockCoverAssetRow, id: 'asset-uuid-2', target: '🏖️' };
-
-    const mockActiveMember = {
-      groupId: 'group-uuid',
-      userId: 'user-uuid',
-      status: GroupMemberStatus.ACTIVE,
-    };
-
-    const mockRequestMember = {
-      groupId: 'group-uuid-2',
-      userId: 'user-uuid',
-      status: GroupMemberStatus.REQUEST,
-    };
-
-    beforeEach(() => {
-      mockAssetResolverResolve.mockResolvedValue(mockResolvedCover);
-    });
-
-    it('returns empty result when no public groups match', async () => {
-      mockGroupMembersFindMany.mockResolvedValueOnce([]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 0 }]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.data).toHaveLength(0);
-      expect(result.total).toBe(0);
-    });
-
-    it('returns only PUBLIC non-deleted groups', async () => {
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([]) // active memberships
-        .mockResolvedValueOnce([]) // active member counts
-        .mockResolvedValueOnce([]); // user membership status
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]); // data
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.total).toBe(1);
-      expect(result.data).toHaveLength(1);
-      expect(result.data[0]?.id).toBe('group-uuid-2');
-    });
-
-    it('excludes groups where user is already an active member', async () => {
-      // User is ACTIVE in group-uuid — it should be excluded from search results
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([mockActiveMember]) // active memberships → exclude group-uuid
-        .mockResolvedValueOnce([]) // active member counts
-        .mockResolvedValueOnce([]); // user membership status
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]); // data
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.data.every((g) => g.id !== 'group-uuid')).toBe(true);
-    });
-
-    it('returns membershipStatus "none" when user has no row', async () => {
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([]) // active memberships
-        .mockResolvedValueOnce([]) // active member counts
-        .mockResolvedValueOnce([]); // user membership status → none
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.data[0]?.membershipStatus).toBe('none');
-    });
-
-    it('returns membershipStatus "pending" when user has a REQUEST row', async () => {
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([]) // active memberships
-        .mockResolvedValueOnce([]) // active member counts
-        .mockResolvedValueOnce([mockRequestMember]); // user membership status → REQUEST
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.data[0]?.membershipStatus).toBe('pending');
-    });
-
-    it('returns membershipStatus "pending" when user has an INVITED row', async () => {
-      const invitedMember = { ...mockRequestMember, status: GroupMemberStatus.INVITED };
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([invitedMember]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.data[0]?.membershipStatus).toBe('pending');
-    });
-
-    it.each([GroupMemberStatus.REMOVED, GroupMemberStatus.LEFT, GroupMemberStatus.REJECTED])(
-      'returns membershipStatus "none" when user has a %s row',
-      async (status) => {
-        const staleMember = { groupId: 'group-uuid-2', userId: 'user-uuid', status };
-        mockGroupMembersFindMany
-          .mockResolvedValueOnce([])
-          .mockResolvedValueOnce([])
-          .mockResolvedValueOnce([staleMember]);
-        mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-        mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-        mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-        const result = await service.searchGroups('user-uuid', baseQuery);
-
-        expect(result.data[0]?.membershipStatus).toBe('none');
-      },
-    );
-
-    it('returns correct memberCount based on active members', async () => {
-      const memberRow1 = { groupId: 'group-uuid-2' };
-      const memberRow2 = { groupId: 'group-uuid-2' };
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([]) // active memberships
-        .mockResolvedValueOnce([memberRow1, memberRow2]) // active member counts → 2
-        .mockResolvedValueOnce([]); // user membership status
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.data[0]?.memberCount).toBe(2);
-    });
-
-    it('returns memberCount 0 when group has no active members', async () => {
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([]) // no active members
-        .mockResolvedValueOnce([]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', baseQuery);
-
-      expect(result.data[0]?.memberCount).toBe(0);
-    });
-
-    it('respects limit and offset pagination', async () => {
-      const paginatedQuery: SearchGroupsQueryDto = { limit: 1, offset: 1 };
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 2 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]); // offset=1 → second group only
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-
-      const result = await service.searchGroups('user-uuid', paginatedQuery);
-
-      expect(result.total).toBe(2);
-      expect(result.data).toHaveLength(1);
-    });
-
-    it('returns empty data array when page is beyond total', async () => {
-      const farQuery: SearchGroupsQueryDto = { limit: 20, offset: 100 };
-      mockGroupMembersFindMany.mockResolvedValueOnce([]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([]); // offset=100 → no results
-
-      const result = await service.searchGroups('user-uuid', farQuery);
-
-      expect(result.total).toBe(1);
-      expect(result.data).toHaveLength(0);
-    });
-
-    it('throws NotFoundException when cover asset is missing for a result', async () => {
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-      mockAssetsFindMany.mockResolvedValue([]); // no asset found
-
-      await expect(service.searchGroups('user-uuid', baseQuery)).rejects.toThrow(NotFoundException);
-    });
-
-    it('throws NotFoundException when cover resolution fails', async () => {
-      mockGroupMembersFindMany
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([])
-        .mockResolvedValueOnce([]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 1 }]);
-      mockGroupsFindMany.mockResolvedValueOnce([mockGroupRow2]);
-      mockAssetsFindMany.mockResolvedValue([mockCoverAssetRow2]);
-      mockAssetResolverResolve.mockResolvedValue(null);
-
-      await expect(service.searchGroups('user-uuid', baseQuery)).rejects.toThrow(NotFoundException);
-    });
-
-    it('uses default limit 20 and offset 0 when not provided', async () => {
-      const queryWithDefaults: SearchGroupsQueryDto = {};
-      mockGroupMembersFindMany.mockResolvedValueOnce([]);
-      mockGroupsSelectWhere.mockResolvedValueOnce([{ total: 0 }]);
-
-      const result = await service.searchGroups('user-uuid', queryWithDefaults);
-
-      expect(result.total).toBe(0);
-      expect(result.data).toHaveLength(0);
     });
   });
 });
